requirements.txt
fastapi==0.104.1
python-dotenv==1.0.0
openai==1.1.0
uvicorn==0.24.0
python-multipart==0.0.6
pymongo==4.6.1
gradio==4.8.0